requirements.txt
Flask
python-dotenv
requests
Flask-WTF
gunicorn